Move functional tests over to functional dir
